MultiAgentLangGraph/requirements.txt
# requirements.txt
langchain-openai>=0.0.2
langchain-community>=0.0.10
langgraph>=0.0.10
requests>=2.31.0
python-dotenv>=1.0.0